move mint to test fixture
